requirements.txt
pytest==7.4.0
requests==2.31.0
PyYAML==6.0.1
allure-pytest==2.13.0
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-ordering==0.6
pytest-rerunfailures==13.0
pytest-timeout==2.2.0
Faker==20.0.1